school-league cleanup: tier 2
Backend:
- Rename _to_csv_export_url to to_csv_export_url. Tests already imported
  it directly, so it was effectively part of the module's API.
- Drop the start= parameter from next_available_team_name and the manual
  counter-bumping in create_school_team's retry. After session.rollback()
  under READ COMMITTED, a competing transaction's commit becomes visible
  to the next query, so re-running next_available_team_name picks the
  next free number on its own.
- The integrity-retry test no longer asserts a "+1" jump on a synthetic
  IntegrityError - that behavior was an artifact of the old start-bump.
  The "skip past taken numbers" semantics are already covered by
  test_next_available_team_name_with_gaps.

Frontend:
- Extract DirectClassicSignup so the parent DirectLeagueSignup is just a
  thin loader that fetches league info and dispatches to the right child.
  Restores symmetry with DirectSchoolLeagueSignup, which was already its
  own component. DirectLeagueSignup goes from 232 to 69 lines.
- Collapse the duplicated body of directSignup and directSchoolSignup in
  useAuthAPI.js into one dispatchSignup helper. Both endpoints return the
  canonical team name in data.team_name, so the auth-state setup is
  identical. Drops the unused leagueInfo parameter from directSignup
  (only one caller, updated in this commit).

Net: -198 lines, 466/466 tests passing.
