+++ backend/routes/user/team_naming.py
@@ -17,8 +17,8 @@ logger = logging.getLogger(__name__)
 MAX_COLLISION_RETRIES = 10
 
 
-def next_available_team_name(session: Session, sanitized: str, start: int = 1) -> str:
-    """Find the lowest N >= start such that f'{sanitized}{N}' is globally unused
+def next_available_team_name(session: Session, sanitized: str) -> str:
+    """Find the lowest N >= 1 such that f'{sanitized}{N}' is globally unused
     as Team.name.
 
     Counter scope: Team.name is globally unique, so this is effectively a
@@ -39,7 +39,7 @@ def next_available_team_name(session: Session, sanitized: str, start: int = 1) -
         if suffix.isdigit():
             used.add(int(suffix))
 
-    n = start
+    n = 1
     while n in used:
         n += 1
     return f"{sanitized}{n}"
@@ -50,8 +50,10 @@ def create_school_team(
 ) -> Team:
     """Create a team for a school league with sanitized name + counter.
 
-    Race-safe: on IntegrityError from the unique-name constraint, bump the
-    counter past the collision and retry up to MAX_COLLISION_RETRIES times.
+    Race-safe: on IntegrityError from the unique-name constraint, roll back
+    and re-query — a competing transaction's commit becomes visible after
+    rollback under READ COMMITTED, so the next candidate naturally skips
+    past it. Retries up to MAX_COLLISION_RETRIES times.
     """
     league = session.get(League, league_id)
     if not league:
@@ -63,9 +65,8 @@ def create_school_team(
             f"School name '{school_name}' contains no alphanumerics"
         )
 
-    start = 1
     for attempt in range(MAX_COLLISION_RETRIES):
-        candidate = next_available_team_name(session, sanitized, start=start)
+        candidate = next_available_team_name(session, sanitized)
         try:
             team = Team(
                 name=candidate,
@@ -84,11 +85,6 @@ def create_school_team(
             logger.warning(
                 "Race on team name %s; retrying (attempt %d)", candidate, attempt + 1
             )
-            try:
-                start = int(candidate[len(sanitized):]) + 1
-            except ValueError:
-                start += 1
-            continue
     raise RuntimeError(
         f"Failed to create school team after {MAX_COLLISION_RETRIES} retries"
     )
